refactor: EventListenerRegistry 분리
